+++ fcrepo-import/src/main/java/org/fcrepo/importer/Importer.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.fcrepo.importexport;
+package org.fcrepo.importer;
 
 /**
  * @author awoods
