+++ fcrepo-import/src/test/java/org/fcrepo/importer/ImportExportDriverTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.fcrepo.importexport;
+package org.fcrepo.importer;
 
 import junit.framework.Test;
 import junit.framework.TestCase;
